backend/app/auth.py
# auth.py - password hashing + JWT util
from datetime import datetime, timedelta
from passlib.context import CryptContext
from jose import JWTError, jwt
import os


SECRET_KEY = os.getenv('SECRET_KEY', 'dev-secret-please-change')
ALGORITHM = 'HS256'
ACCESS_TOKEN_EXPIRE_MINUTES = 60*24


pwd_context = CryptContext(schemes=['bcrypt'], deprecated='auto')


# simple helpers
def hash_password(password: str) -> str:
return pwd_context.hash(password)


def verify_password(plain: str, hashed: str) -> bool:
return pwd_context.verify(plain, hashed)


def create_access_token(data: dict, expires_delta: timedelta | None = None):
to_encode = data.copy()
expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
to_encode.update({"exp": expire})
return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)


def decode_token(token: str):
try:
payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
return payload
except JWTError:
return None