backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
passlib[bcrypt]
python-jose
python-multipart
alembic
python-dotenv